requirements.txt
discord.py==2.1.0
discord.py-ipc==1.1.0
quart==0.18.3
Quart-Discord==2.1.4